# perf: `ape_node` plugin load time improvement