### What I did

noooooo i wish this was in the last release, so close to have a faster default Ape ecosystem..

(ipython)

before

```
In [1]: %time import ape_node
CPU times: user 1.14 s, sys: 142 ms, total: 1.28 s
Wall time: 1.29 s
```

after:

```
In [1]: %time import ape_node
CPU times: user 7.93 ms, sys: 3.39 ms, total: 11.3 ms
Wall time: 10 ms
```

BASICALLY once all our plugins have this treatment, stuff like `ape console` will be able to launch faster, if my research is correct. And i already have ideas of where the bottleneck moves to. We're optimizing.

### How I did it

<!-- Discuss the thought process behind the change -->

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated